+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-03 09:51:46
-// Commit: fix(theme): prevent double tap duplicate action in theme
+// Updated at: 2026-08-03 16:09:05
+// Commit: feat(theme): add bi-directional event bus handler for theme
 
 class themeComponentState {
-  final String commitRef = '3291';
-  final String lastMessage = 'fix(theme): prevent double tap duplicate action in theme';
-  final DateTime timestamp = DateTime.parse('2026-08-03T09:51:46+05:00');
+  final String commitRef = '3305';
+  final String lastMessage = 'feat(theme): add bi-directional event bus handler for theme';
+  final DateTime timestamp = DateTime.parse('2026-08-03T16:09:05+05:00');
 }